Move core to a subfolder and leave preact as the main
