+++ package.json
@@ -2,12 +2,13 @@
 	"name": "deepsignals",
 	"private": true,
 	"scripts": {
-		"prebuild": "rimraf packages/deepsignal/dist/ && rimraf packages/deepsignal/preact/dist",
-		"build": "pnpm build:deepsignal && pnpm build:deepsignal-preact",
-		"build:deepsignal": "microbundle --raw --cwd packages/deepsignal",
-		"build:deepsignal-preact": "microbundle --raw --cwd packages/deepsignal/preact && pnpm postbuild:deepsignal-preact",
+		"prebuild": "rimraf packages/deepsignal/dist/ && rimraf packages/deepsignal/core/dist",
+		"build": "pnpm build:deepsignal && pnpm build:deepsignal-core",
+		"build:deepsignal": "microbundle --raw --cwd packages/deepsignal && pnpm postbuild:deepsignal",
+		"build:deepsignal-core": "microbundle --raw --cwd packages/deepsignal/core && pnpm postbuild:deepsignal-core",
 		"postbuild": "node ./scripts/node-13-exports.js",
-		"postbuild:deepsignal-preact": "cd packages/deepsignal/preact/dist && mv -f src/index.d.ts deepsignal-preact.d.ts && rm -dr preact && rm -dr src",
+		"postbuild:deepsignal": "cd packages/deepsignal/dist && mv -f core/src/index.d.ts deepsignal.d.ts && rm -dr core && rm -dr src",
+		"postbuild:deepsignal-core": "cd packages/deepsignal/core/dist && mv -f index.d.ts deepsignal-core.d.ts",
 		"lint": "eslint 'packages/**/*.{ts,tsx,js,jsx}'",
 		"test": "cross-env COVERAGE=true karma start karma.conf.js --single-run",
 		"test:minify": "cross-env COVERAGE=true MINIFY=true karma start karma.conf.js --single-run",
