--- packages/deepsignal/preact/src/index.ts
@@ -1,9 +1,0 @@
-import "@preact/signals";
-import { useMemo } from "preact/hooks";
-import { deepSignal, type DeepSignal } from "../../src";
-
-export const useDeepSignal = <T extends object>(obj: T): DeepSignal<T> => {
-	return useMemo(() => deepSignal(obj), []);
-};
-
-export { deepSignal, type DeepSignal } from "../../src";
