+++ packages/deepsignal/src/index.ts
@@ -1,142 +1,9 @@
-import { computed, signal, Signal } from "@preact/signals-core";
+import "@preact/signals";
+import { useMemo } from "preact/hooks";
+import { deepSignal, type DeepSignal } from "../core/src";
 
-const proxyToSignals = new WeakMap();
-const objToProxy = new WeakMap();
-const arrayToArrayOfSignals = new WeakMap();
-const rg = /^\$\$?/;
-
-type DeepSignalObject<T extends object> = {
-	[P in keyof T & string as `$${P}`]?: Signal<T[P]>;
-} & {
-	[P in keyof T & string as `$$${P}`]?: T[P];
-} & {
-	[P in keyof T]: T[P] extends Array<unknown>
-		? DeepSignalArray<T[P]>
-		: T[P] extends object
-		? DeepSignalObject<T[P]>
-		: T[P];
+export const useDeepSignal = <T extends object>(obj: T): DeepSignal<T> => {
+	return useMemo(() => deepSignal(obj), []);
 };
 
-type ArrayType<T> = T extends Array<infer I> ? I : T;
-type DeepSignalArray<T> = Array<ArrayType<T>> & {
-	[key: number]: DeepSignal<ArrayType<T>>;
-	$?: { [key: number]: Signal<ArrayType<T>> };
-	$$?: { [key: number]: ArrayType<T> };
-	$length?: Signal<number>;
-	$$length?: number;
-};
-
-export type DeepSignal<T> = T extends Array<unknown>
-	? DeepSignalArray<T>
-	: T extends object
-	? DeepSignalObject<T>
-	: T;
-
-export const deepSignal = <T extends object>(obj: T): DeepSignal<T> => {
-	if (!shouldProxy(obj)) throw new Error("This object can't be observed.");
-	if (!objToProxy.has(obj))
-		objToProxy.set(obj, new Proxy(obj, objectHandlers) as DeepSignal<T>);
-	return objToProxy.get(obj);
-};
-
-const get =
-	(isArrayOfSignals: boolean) =>
-	(target: object, fullKey: string, receiver: object): unknown => {
-		let returnSignal = isArrayOfSignals || fullKey[0] === "$";
-		if (!isArrayOfSignals && returnSignal && Array.isArray(target)) {
-			if (fullKey === "$") {
-				if (!arrayToArrayOfSignals.has(target))
-					arrayToArrayOfSignals.set(target, new Proxy(target, arrayHandlers));
-				return arrayToArrayOfSignals.get(target);
-			}
-			if (fullKey === "$$") return target;
-			returnSignal = fullKey === "$length" || fullKey === "$$length";
-		}
-		if (!proxyToSignals.has(receiver)) proxyToSignals.set(receiver, new Map());
-		const signals = proxyToSignals.get(receiver);
-		const key = returnSignal ? fullKey.replace(rg, "") : fullKey;
-		if (
-			!signals.has(key) &&
-			typeof Object.getOwnPropertyDescriptor(target, key)?.get === "function"
-		) {
-			if (returnSignal && fullKey[1] === "$")
-				return Reflect.get(target, key, receiver);
-			signals.set(
-				key,
-				computed(() => Reflect.get(target, key, receiver))
-			);
-		} else {
-			let value = Reflect.get(target, key, receiver);
-			if (typeof key === "symbol" && wellKnownSymbols.has(key)) return value;
-			if (returnSignal && fullKey[1] === "$") return value;
-			if (!signals.has(key)) {
-				if (shouldProxy(value)) {
-					if (!objToProxy.has(value))
-						objToProxy.set(value, new Proxy(value, objectHandlers));
-					value = objToProxy.get(value);
-				}
-				signals.set(key, signal(value));
-			}
-		}
-		return returnSignal ? signals.get(key) : signals.get(key).value;
-	};
-
-const mutationError = "Don't mutate the signals directly.";
-
-const objectHandlers = {
-	get: get(false),
-	set(target: object, key: string, val: any, receiver: object) {
-		if (key[0] === "$") throw new Error(mutationError);
-		let internal = val;
-		if (shouldProxy(val)) {
-			if (!objToProxy.has(val))
-				objToProxy.set(val, new Proxy(val, objectHandlers));
-			internal = objToProxy.get(val);
-		}
-		if (!proxyToSignals.has(receiver)) proxyToSignals.set(receiver, new Map());
-		const signals = proxyToSignals.get(receiver);
-		if (!signals.has(key)) signals.set(key, signal(internal));
-		else signals.get(key).value = internal;
-		const result = Reflect.set(target, key, val, receiver);
-		if (Array.isArray(target) && signals.has("length"))
-			signals.get("length").value = target.length;
-		return result;
-	},
-};
-
-const arrayHandlers = {
-	get: get(true),
-	set() {
-		throw new Error(mutationError);
-	},
-};
-
-type WellKnownSymbols =
-	| "asyncIterator"
-	| "hasInstance"
-	| "isConcatSpreadable"
-	| "iterator"
-	| "match"
-	| "matchAll"
-	| "replace"
-	| "search"
-	| "species"
-	| "split"
-	| "toPrimitive"
-	| "toStringTag"
-	| "unscopables";
-
-const wellKnownSymbols = new Set(
-	Object.getOwnPropertyNames(Symbol)
-		.map(key => Symbol[key as WellKnownSymbols])
-		.filter(value => typeof value === "symbol")
-);
-const supported = new Set([Object, Array]);
-const shouldProxy = (val: any): boolean => {
-	if (typeof val !== "object" || val === null) return false;
-	const isBuiltIn =
-		typeof val.constructor === "function" &&
-		val.constructor.name in globalThis &&
-		(globalThis as any)[val.constructor.name] === val.constructor;
-	return !isBuiltIn || supported.has(val.constructor);
-};
+export { deepSignal, type DeepSignal } from "../core/src";
