+++ tsconfig.json
@@ -13,7 +13,7 @@
 		"noUnusedLocals": true,
 		"paths": {
 			"deepsignal": ["./packages/deepsignal/src/index.ts"],
-			"deepsignal/preact": ["./packages/deepsignal/preact/src/index.ts"]
+			"deepsignal/core": ["./packages/deepsignal/core/src/index.ts"]
 		}
 	}
 }
